+++ spec/faraday/detailed_logger/middleware_spec.rb
@@ -45,12 +45,36 @@ CURL
     expect(log.read).to match(/\bDEBUG\b.+#{Regexp.escape(curl.inspect)}/)
   end
 
-  it "logs the response status code at an INFO level" do
+  it "logs a 2XX response status code at an INFO level" do
     logger = Logger.new(log = StringIO.new)
 
-    connection(logger).get("/temaki")
+    connection(logger).get("/oaiso", {c: 200})
     log.rewind
     expect(log.read).to match(/\bINFO\b.+\bHTTP 200\b/)
+  end
+
+  it "logs a 3XX response status code at an INFO level" do
+    logger = Logger.new(log = StringIO.new)
+
+    connection(logger).get("/oaiso", {c: 301})
+    log.rewind
+    expect(log.read).to match(/\bINFO\b.+\bHTTP 301\b/)
+  end
+
+  it "logs a 4XX response status code at a WARN level" do
+    logger = Logger.new(log = StringIO.new)
+
+    connection(logger).get("/oaiso", {c: 401})
+    log.rewind
+    expect(log.read).to match(/\bWARN\b.+\bHTTP 401\b/)
+  end
+
+  it "logs a 5XX response status code at an WARN level" do
+    logger = Logger.new(log = StringIO.new)
+
+    connection(logger).get("/oaiso", {c: 500})
+    log.rewind
+    expect(log.read).to match(/\bWARN\b.+\bHTTP 500\b/)
   end
 
   it "logs a cURL-like response package at a DEBUG level" do
@@ -80,6 +104,9 @@ CURL
         }
         stub.post("/nigirizushi") {
           [200, {"Content-Type" => "application/json"}, "{\"order_id\":\"1\"}"]
+        }
+        stub.get("/oaiso") { |env|
+          [env.params["c"].to_i, {"Content-Type" => "application/json"}, env.params["c"]]
         }
       end
     end